+++ src/spellbot/actions/admin_action.py
@@ -220,6 +220,15 @@ class AdminAction(BaseAction):
         embeds = await self._build_awards_embeds()
         await safe_send_channel(self.interaction, embed=embeds[page - 1])
 
+    async def transfer(
+        self,
+        old: discord.Member,
+        new: discord.Member,
+        guild: discord.Guild,
+    ) -> None:
+        await self.services.games.transfer(old.id, new.id, guild.id)
+        await safe_send_channel(self.interaction, "Done!")
+
     async def award_add(
         self,
         count: int,